# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewStateRead": {
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "clean_count": {
            "title": "Clean Count",
            "type": "integer"
          },
          "due_at": {
            "format": "date-time",
            "title": "Due At",
            "type": "string"
          },
          "mastery_stage": {
            "title": "Mastery Stage",
            "type": "string"
          },
          "segment_id": {
            "title": "Segment Id",
            "type": "string"
          }
        },
        "required": [
          "segment_id",
          "due_at",
          "mastery_stage",
          "clean_count",
          "attempt_count"
        ],
        "title": "ReviewStateRead",
        "type": "object"
      },
...
      "WeakLinkRead": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "difficult_attempts": {
            "title": "Difficult Attempts",
            "type": "integer"
          },
          "difficulty_rate": {
            "title": "Difficulty Rate",
            "type": "number"
          },
          "segment_id": {
            "title": "Segment Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "segment_id",
          "text",
          "attempts",
          "difficult_attempts",
          "difficulty_rate"
        ],
        "title": "WeakLinkRead",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local-first API for structured oral memorization.",
    "title": "Rhapsode API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analytics/due": {
      "get": {
        "operationId": "due_reviews_api_v1_analytics_due_get",
        "parameters": [
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewStateRead"
                  },
                  "title": "Response Due Reviews Api V1 Analytics Due Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Due Reviews",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/weak-links": {
      "get": {
        "operationId": "weak_links_api_v1_analytics_weak_links_get",
        "parameters": [
          {
            "in": "query",
            "name": "revision_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WeakLinkRead"
                  },
                  "title": "Response Weak Links Api V1 Analytics Weak Links Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Weak Links",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/annotations": {
      "post": {
        "operationId": "create_annotation_api_v1_annotations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Annotation",
        "tags": [
          "passages"
        ]
      }
    },
    "/api/v1/annotations/{annotation_id}": {
      "delete": {
        "operationId": "delete_annotation_api_v1_annotations__annotation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "annotation_id",
            "required": true,
            "schema": {
              "title": "Annotation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Annotation Api V1 Annotations  Annotation Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Annotation",
        "tags": [
          "passages"
        ]
      }
    },
...
    "/api/v1/languages": {
      "get": {
        "operationId": "list_languages_api_v1_languages_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LanguageProfileRead"
                  },
                  "title": "Response List Languages Api V1 Languages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Languages",
        "tags": [
          "languages"
        ]
      },
      "post": {
        "operationId": "create_language_api_v1_languages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanguageProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageProfileRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Language",
        "tags": [
          "languages"
        ]
      }
    },
    "/api/v1/media": {
      "post": {
        "operationId": "upload_media_api_v1_media_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_media_api_v1_media_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Media",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/{media_id}": {
      "delete": {
        "operationId": "delete_media_api_v1_media__media_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "media_id",
            "required": true,
            "schema": {
              "title": "Media Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Media Api V1 Media  Media Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Media",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/{media_id}/content": {
      "get": {
        "operationId": "stream_media_api_v1_media__media_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "media_id",
            "required": true,
            "schema": {
              "title": "Media Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Media",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/passages": {
      "get": {
        "operationId": "list_passages_api_v1_passages_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PassageRead"
                  },
                  "title": "Response List Passages Api V1 Passages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Passages",
        "tags": [
          "passages"
        ]
      },
      "post": {
        "operationId": "create_passage_api_v1_passages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PassageInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassageDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Passage",
        "tags": [
          "passages"
        ]
      }
    },
    "/api/v1/passages/{passage_id}": {
      "get": {
        "operationId": "get_passage_api_v1_passages__passage_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "passage_id",
            "required": true,
            "schema": {
              "title": "Passage Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassageDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Passage",
        "tags": [
          "passages"
        ]
      }
    },
    "/api/v1/passages/{passage_id}/revisions": {
      "post": {
        "operationId": "create_revision_api_v1_passages__passage_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "passage_id",
            "required": true,
            "schema": {
              "title": "Passage Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Revision",
        "tags": [
          "passages"
        ]
      }
    },
    "/api/v1/plugins": {
      "get": {
        "operationId": "list_plugins_api_v1_plugins_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PluginRead"
                  },
                  "title": "Response List Plugins Api V1 Plugins Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Plugins",
        "tags": [
          "plugins"
        ]
      },
      "post": {
        "operationId": "register_plugin_api_v1_plugins_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Plugin",
        "tags": [
          "plugins"
        ]
      }
    },
    "/api/v1/revisions/{revision_id}": {
      "get": {
        "operationId": "get_revision_api_v1_revisions__revision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Revision",
        "tags": [
          "passages"
        ]
      }
    },
...
  }
}